+++ lib/index.js
@@ -79,9 +79,13 @@ function Api(options) {
 	api.express.use(bodyParser.json(options['body-parser']));
 
 	// DEFAULT LISTENERS
-	api.chainListener('connection', require('./verify-token')(url, jwt));
-	api.chainListener('message', require('./add-message-context')(_));
-	api.chainListener('request', require('./add-request-context')());
+	var verifyToken = require('./listeners/verify-token');
+	var addMessageContext = require('./listeners/add-message-context');
+	var addRequestContext = require('./listeners/add-request-context');
+
+	api.chainListener('connection', verifyToken(url, jwt));
+	api.chainListener('message', addMessageContext(_));
+	api.chainListener('request', addRequestContext());
 
 	// CUSTOM LISTENERS
 	var addConnectionHooks = _.bind(api.chainListener, api, 'connection');
