+++ lib/listeners/verify-token.js
